+++ .gitignore
@@ -8,6 +8,7 @@
 .buildlog/
 .history
 .svn/
+coverage/
 
 # IntelliJ related
 *.iml
